obj_lv5ychndlr done

+++ include/d/a/d_a_player.h
@@ -674,7 +674,7 @@ public:
     virtual bool checkSpinnerPathMove();
     virtual bool checkSpinnerTriggerAttack();
     virtual void onSpinnerPathForceRemove();
-    virtual s16 getIronBallBgHit() const;
+    virtual int getIronBallBgHit() const;
     virtual cXyz* getIronBallCenterPos();
     virtual bool checkCanoeFishingGetLeft() const;
     virtual bool checkCanoeFishingGetRight() const;
